handle launch qwith a real task to poll openhexa
