+++ hat/assets/js/apps/Iaso/domains/assignments/sampling/OpenhexaIntegrationDrawer.tsx
@@ -33,7 +33,6 @@ import { TaskLogMessages } from 'Iaso/domains/tasks/components/TaskLogMessages';
 import { useGetLogs } from 'Iaso/domains/tasks/hooks/api';
 
 import { SxStyles } from 'Iaso/types/general';
-import { usePollTask } from '../hooks/requests/usePollTask';
 import MESSAGES from '../messages';
 import { Planning } from '../types/planning';
 import { StatusInfos } from './StatusInfos';
@@ -121,15 +120,29 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
         isLoading: isLaunchingTask,
     } = useLaunchTask(selectedPipelineId, pipeline?.currentVersion?.id, false);
     const taskId = launchResult?.task?.id;
+    const [taskStatus, setTaskStatus] = useState<string | undefined>(undefined);
+
+    const isPipelineActive =
+        (!taskStatus || taskStatus === 'RUNNING' || taskStatus === 'QUEUED') &&
+        currentStep === 2;
     const handleEndTask = useCallback(() => {
         queryClient.invalidateQueries(['assignmentsList']);
     }, [queryClient]);
-    const { data: task } = usePollTask(taskId, handleEndTask);
+
     const { data: taskLogs, isFetching: isFetchingTaskLogs } = useGetLogs(
         taskId,
-        task?.status === 'RUNNING',
+        isPipelineActive,
     );
 
+    useEffect(() => {
+        if (taskLogs && taskLogs.status !== taskStatus) {
+            setTaskStatus(taskLogs.status);
+            if (taskLogs.status === 'SUCCESS') {
+                handleEndTask();
+            }
+        }
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [taskLogs]);
     const pipelineUuidsOptions = useMemo(
         () =>
             data?.filter(pipeline =>
@@ -166,8 +179,6 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
             setIsSubmitting(false);
         }
     }, [isSubmitting, isLaunchingTask]);
-
-    const isPipelineRunning = task?.status === 'RUNNING';
     return (
         <>
             <Tooltip title={disabled ? disabledMessage : undefined}>
@@ -237,7 +248,7 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
                                         onClick={handleSubmit}
                                         disabled={
                                             !allowConfirm ||
-                                            isPipelineRunning ||
+                                            isPipelineActive ||
                                             isFetchingPipeline
                                         }
                                         fullWidth
@@ -251,7 +262,7 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
                                         variant="contained"
                                         onClick={() => setCurrentStep(1)}
                                         disabled={
-                                            isPipelineRunning || isLaunchingTask
+                                            isPipelineActive || isLaunchingTask
                                         }
                                         fullWidth
                                     >
@@ -263,7 +274,7 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
                         </Grid>
                     </Paper>
                     <Box position="relative" sx={styles.container}>
-                        {(isFetchingPipeline || isPipelineRunning) && (
+                        {(isFetchingPipeline || isPipelineActive) && (
                             <LoadingSpinner absolute fixed={false} />
                         )}
                         {/* Step 1: Pipeline selection and parameters */}
@@ -339,7 +350,9 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
                                         message={error.details.error}
                                     />
                                 )}
-                                {task && <StatusInfos status={task.status} />}
+                                {taskLogs && (
+                                    <StatusInfos status={taskLogs.status} />
+                                )}
                                 {taskId &&
                                     taskLogs &&
                                     taskLogs?.logs?.length > 0 && (
@@ -347,7 +360,8 @@ export const OpenhexaIntegrationDrawer: FunctionComponent<Props> = ({
                                             <TaskLogMessages
                                                 messages={taskLogs.logs}
                                                 isFetching={isFetchingTaskLogs}
-                                                isRunning={isPipelineRunning}
+                                                isRunning={isPipelineActive}
+                                                displayLoader={false}
                                             />
                                         </Box>
                                     )}
